adding util.py and check in harmonise_data.py to see if build mapping is needed
