+++ final_formatting/utils.py
@@ -1,0 +1,13 @@
+import csv
+
+from common_constants import *
+
+
+def get_csv_reader(csv_file):
+    dialect = csv.Sniffer().sniff(csv_file.readline())
+    csv_file.seek(0)
+    return csv.reader(csv_file, dialect)
+
+
+def get_filename(file):
+    return file.split("/")[-1].split(".")[0]
